NIGGER NIGGER NIGGER NIGGER NIGGER NIGGER

+++ Labo2_SMI/src/delai.h
@@ -3,7 +3,7 @@
 #ifndef DELAI_H_
 #define DELAI_H_
 
-void SysTick_Init(uint64_t load_value);
+void SysTick_Init(uint32_t load_value);
 
 uint64_t millis();
